[FIX] l10n_mx_edi: Show the 'Update SAT' button when cancel requested
At the moment, if a cancellation has been requested, the 'Update SAT'
button isn't always visible on the invoice.

Analysis:
The condition for the 'Update SAT' button being visible is that at
least one of the EDI documents falls under one of the domains in
`_get_update_sat_status_domains`. In particular, the 'Update SAT' status
is visible if there is an EDI document with the 'invoice_sent' status
and whose SAT status is not 'cancelled' or 'skip'.

If the user requests the cancellation of the CFDI before the SAT status
is fetched for the first EDI document (which represents the sending of
the invoice), its SAT status will be set to 'skip'.

As a result, the 'Update SAT' button will no longer be visible, meaning
that the user can not update the invoice's state from fetching the SAT
status.

Fix:
The Update SAT button should also be visible if there is a document in
'invoice_cancel_requested' whose SAT state is valid.

opw-4375788

closes odoo/enterprise#83399

X-original-commit: 872bc60506dc74c66c2f189ae59d26e562a540cc
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>