[FIX] l10n_it_xml_export: missing options in closing entry
Description of the issue this commit addresses:

When posting the closing entry of an Italian tax return, with the
l10n_it_xml_export module installed, a traceback is raised.

Steps to reproduce:

1. Install l10n_it_xml_export.
2. Using an IT company, create a closing entry for the tax return.
3. Post the closing entry, a wizard opens.
4. Validate the values.
5. A traceback is raised.

Desired behavior after this commit is merged:

The export works and the xml file is downloaded.

opw-4715039

closes odoo/enterprise#83342

X-original-commit: 7323435468c8db15cd5643f7d6ecf5928ab2f3dc
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>