[IMP] hr_appraisal{,_skills}: introduction of goal template
In this commit, we introduce a new feature in hr_appraisal : goal template.
With this new feature it's possible to save a goal and to reuse it for another employee.
A goal template can also have sub goal templates; really useful to create more complex goals.

task-4680306

closes odoo/enterprise#83658

Related: odoo/odoo#225129
Related: odoo/upgrade#8390
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>