[IMP] l10n_br_{avatax,avatax_sale,edi,edi_pos}: support different operation types per line
Before, the operation types could only be set on the Sale Order or on
the Invoice level. This works for most cases but if a customer needs to
invoice products by different operation types they can't.

For example, if a customer is selling some products that come with a
free sample, the sample will be a different operation type than the
normal "sale of goods". Either we can create two different invoices /
sale orders for the different items (which is legally not correct) or we
can create multiple products per operation type which is complex.

Therefore we need to add a new field to account.move.line /
sale.order.line

This commit makes it possible to pick from the same list of values
as on the parent orders/invoices as an "override" value for an specific
lines that need a different operation type. This means that for any line
that does not have an override set, the field of the parent model is
used instead.

For service products, Avatax still expects the operations type to be set
in the header. As such we must still send it, but for any record that
has a different tax in the lines versus the header, it will prioritize
the lines value.

task-4569548

closes odoo/enterprise#81525

Related: odoo/upgrade#7390
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>