[IMP] account_{followup|reports}: Customer Statement button on partner form view visibility
Description of the issue this commits addresses:

The Customer Statement button is shown too often and leads to empty customer
statements in some scenarios.

The banner warning the user that some entries are draft is also shown too often
and should only appear if the selected partner(s) have draft entries.

Desired behavior after this commit is merged:

The button only shows when the customer statement it will lead to has values
to show. This means both making sure the partner has moves to show on the
customer statement but also that the customer statement actually shows them.

Same for the draft entries warning banner. It will only be shown when the
selected parter(s) has(have) draft entries.

task-4680249

closes odoo/enterprise#83346

X-original-commit: f9cf8550a38527379520765f767f2043f42df78d
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>