[FIX] delivery_sendcloud: avoid setting to_service_point on return label
Issue:

It is currently not possible to generate a return label with sendcloud
if the delivery has been generated using a service point.

Step to reproduce:

- Set up Sendcloud, for a belgian company
- Select bpost and dpd for delivery and returns while loading products
- On a private window buy a storable product on the webstore online and
  select a sendcloud pickup point as the delivery method.
- With an employee confirm the delivery order that was created
> A problem occured in the generation of the return label

Cause of the issue:

The `to_service_point` parameter is used by sendcloud's API in order to
generate parcels and is added here:
https://github.com/odoo/enterprise/blob/747482f34ca8210d89d81b3363ec5a8fc3a4ff6f/delivery_sendcloud/models/sendcloud_service.py#L109-L112
However, this parameter does not make sense in case of a return since
the customer is allowed to return the product to the pickup point of its
choice. In particular, sendcloud's API will fail to genrate the return
parcel if it is expecting to link the label to a given pickup point.

Note:

Sendcloud also fails to generate the return label when the shipping
provider is the same for the delivery and for the return, so that
the `to_service_point` can just not be provided for return labels.

opw-4562040

closes odoo/enterprise#83902

X-original-commit: 5735e06d13991772a5f0c16a19bbb577ea74c08e
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>