[FIX] appointment: prevent error while clicking on Add Closing Day(s)
When timezone of user is empty and the user clicks on Add Closing Day(s) button,
a traceback will appear.

Steps to reproduce the error:
- Open Profile > Preferences > Select empty in timezone > Save
- Install 'Appointments'
- Go to Appointments > Schedule > Resource Bookings > Add Closing Day(s)

Traceback:
"AttributeError: 'bool' object has no attribute 'upper'"

https://github.com/odoo/enterprise/blob/2c938b66fc1b526a360f4b199a47bcc53991bcb9/appointment/wizard/appointment_manage_leaves.py#L14
When the user selects an empty timezone, tz will be False.
So, It will lead to the above traceback.

Forward port https://github.com/odoo/enterprise/pull/64948 was closed in master,
so the issue is still present in saas-17.4 and all later versions.

sentry-6537084290

closes odoo/enterprise#85689

X-original-commit: 4a98f90a565118cd158002da9b2f4719dbcf25f1
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>
Signed-off-by: Renilkumar Kajavadra (reka) <reka@odoo.com>