[IMP] documents: make actions available in SHARED
It can be useful that users can execute embedded actions on documents
even when they have no access to the folder itself.

Flows similar to this are then enabled:
A user is shared a document (as viewer) and is expected to
1. Do something about it,
2. Use an embedded action to add a "Done"/"Acknowledged"/"Paid" tag.

Also, cleaning deprecated `action_folder_embed_action` `groups_ids` parameter.

Task-4724349

Part-of: odoo/enterprise#83368
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>