[FIX] sale_amazon: consider SKU containing a comma
The SP-API doesn't support SKU containing commas, not even when escaping
or listing them in a correct CSV like string, as they'll be considered
as either separate SKU or a different one, thus leading to an error
saying there are too many identifiers, and not recognizing them anyway.
When an item isn't recognize, Amazon will just not send any information.

In order to let the synchronization continue for the other offers, ensure
their SKU don't contain commas as it'd be considered as more than one SKU,
and fallback on a FBA like data, so as to not try to fetch this information
everytime.

As from Amazon side, using a comma is a valid value, we can't constrain
the SKU either.

opw-4710435

X-original-commit: 1f53ca45a96b9c8b37c6eaab550f65cd62f62efb
Part-of: odoo/enterprise#83543
Signed-off-by: Morgane Demesmaeker <edm@odoo.com>