[FIX] stock_barcode: all quants accessible in barcode with packages
**Problem:**
When having multiple quants for a product with the same lot but
different packages in the same location. Not all quants are visible
in the barcode app inventory.

**Steps to reproduce:**
- In settings activate the following settings : Packages,
Storage Locations, Count Entire Locations
- Open Inventory/Products
- Create a new product
- Check the "track inventory" field and select "By Lot"
- Go to Inventory/Operations/Physical Inventory and create a new
line for your product
- As Location put WH/Stock/Shelf 1
- Write Lot1 for the Lot and Pack1 for the Package
- Set a quantity of 5
- Save and Apply
- Do the same for another line put but Pack2 for the Package
- Open the barcode app and select inventory adjustments
- Click on the barcode icon on the top right of the screen
- Write SHELF1 and click on apply

**Current behavior:**
Only 5 unit of your product are displayed (those for Pack1)

**Expected behavior:**
10 units should be displayed. The 5 from Pack1 and the
5 from Pack2

**Cause of the issue:**
When checking if the line already exists, the package is not
taken into account
https://github.com/odoo/enterprise/blob/7a132639d6c05ef96d43a48ea34ff11711539fc7/stock_barcode/static/src/models/barcode_quant_model.js#L563-L564
so if a line already exists for the same
lot "!currentLine" will be false and the code inside the if
statement will not be run
https://github.com/odoo/enterprise/blob/7a132639d6c05ef96d43a48ea34ff11711539fc7/stock_barcode/static/src/models/barcode_quant_model.js#L565
resulting in the new line not being added in the barcode
app

opw-4640303

closes odoo/enterprise#85193

X-original-commit: d2bc25f24fcf0fbc5a1362e8428f6e02cc55de86
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Pierre-Louis Hance (plha) <plha@odoo.com>