[IMP] AI: improve ChatGPT integration for messages and HTML fields
In this commit we create the new ai_apps module and improve the chatgpt
integration that already existed in odoo.

Previously, our AI integration but it was pretty naive. It opened in a
new dialog (making the rest of the UI unusable while using it), and with
0 information regarding the record it's being used for.

Now, the chatGPT has been fully integrated in discuss, so every time a
user wants to chat with the AI and generate text - they will chat as if
it's another user. We now also share record information through the AI's
context as well as chatter information when using the AI to compose new
messages. Additionally a new AI button over the chatter was added which
allows the user to summarize the chatter conversation as well as write new
messages and notes.

A lot of methods for styling, new actions, and action conditions were
created / extended for the discuss integration. Also a new ai_composer
model was added, which represents the different composer types that the
AI can have depending on where it's called from.

Finally, we here added the chatgpt plugin that we removed from the html_editor
module.

X-original-commit: 3f866ab3cf01045029d55bc321d9bd2790357fbd
Part-of: odoo/enterprise#85028
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>