[FIX] *external_tax: rename tax group containing total taxes
If you have multiple taxes on a sale order or invoice, then the total
tax amount is displayed with the name of the first tax group. Rename
it to show a generic "Taxes" group.

This was done before for sale orders but was lost in
d5da2b62263f85fbce0f2548085c066cbbf10371.

opw-4673667

closes odoo/enterprise#83408

X-original-commit: 0997472db1a7c640bfbf8109b281f6d7c767f63f
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>