[FIX] timesheet_grid: avoid removing project from timesheet line
Currently it's possible for a timesheet user to set and write the project to
False in the timesheet list view. The UI will say the project is missing but the
analytic line is changed anyway.

This PR add check that the project is set before writing to the analytic line
record.

opw-4649817

closes odoo/enterprise#84928

X-original-commit: d41c538490ce2fcf03ac4f15c67452104effcd73
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Thomas Beckers (tbs) <tbs@odoo.com>