[FIX] sale_timesheet_enterprise_holidays: fix test ACLs
https://runbot.odoo.com/odoo/error/163099

This issue became visible as part of the change to test ACLs (removal
of demo data, possibly change to groups too) as well as running the
"single app tests" on *every module*, not just the apps.

The issue is that `self.Requests.user` and `self.employee_user` are
unrelated (the amazingly named `self.employee_user` doesn't even have
a user). On leave creation, this triggers the rule
`hr_leave_rule_employee_update` which checks if:

- The current user is the employee's, which is not the case because
  the employee doesn't have a user.
- Or the current user is the employee's leave manager, which is also
  not the case because the employee does not have a leaves manager.

As a result the creation of the leaves fails immediately.

Fix in 17.4 as the test was introduced in #53940 which was merged in
(what would become) 17.3, even though the test setup means it might
not be possible to trigger there because of all the demo messing with
groups.

closes odoo/enterprise#83474

X-original-commit: d4fe4b355703e335038d32df7b0328615aaf0ccd
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>