[IMP] planning,web(site)_studio: show control panel asap
This commit is the counter-part of odoo/odoo#205129, where list,
kanban, graph and pivot views have been tweaked s.t. their
control panel is shown directly, without waiting for the data to
be ready.

This commit disables the feature in custom list and kanban views
in planning, because those views customize the way data is fetch
(to fetch additional data), and this doesn't map very well with
the strategy implemented in odoo/odoo#205129.

This commit also fixes the x2many edition in studio, where a new
type of root datapoint can be spawned (StaticList), and a tour
that must now explicitely wait for the data to be ready.

task-4282605

Part-of: odoo/enterprise#83321
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>