[FIX] sale_commission: account for down payments in achievement report
In the sale achievement report, down payments and other account move
lines that have no product attached to them would not be accounted for
in the sale commission achievement report. This was due to the fact that
the table was being inner joined on product_id which would get rid of
records that had no product.

Because of this behavior there would be a mismatch between the invoice
analysis of a user and their commission even if the rate was 100% on the
commission plan.

Left joining the products onto the account move line keeps the lines
with no products and fixes the discrepancy between the report and the
invoice analysis.

opw-4625340

closes odoo/enterprise#91179

X-original-commit: 29c22eaffa6cd885cfbbd67939be28da7de519b3
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Ryan Cen (ryce) <ryce@odoo.com>