[FIX] partner_commission: fix traceback when creating payment with referrar
Currently a traceback is occurring when the user tries to create an
invoice payment with referrer in invoice.

To reproduce this issue:

1) Install partner_commission without demo
2) Create a commission plan from crm with a rule containing category and rate
3) Create a product with the same category used in the above commission rule
4) Create a contact with commission plan and partner level in partner assignment
5) Create a confirmed customer invoice with line as above product and referrer as
   above contact and register a payment.

Error:- "AttributeError: 'bool' object has no attribute 'partner_id'"

The above traceback is occurring because we are trying to get the partner_id
value from `_get_sales_representative` method.
https://github.com/odoo/enterprise/blob/313593cf176271060878dda5ebc677fa44b519db/partner_commission/models/account_move.py#L124-L125

But the above method tries to get the user_id from the sale order.
Here in this case, we don't have any SO, so we the above method will return False.

https://github.com/odoo/enterprise/blob/313593cf176271060878dda5ebc677fa44b519db/partner_commission/models/account_move.py#L26-L27

This leads to the above traceback when parter_id is retrieved from the method.

Solution:-

Take the value of partner_id from purchase. Since the partner_id value is getting
from move.referrer_id in purchase.

https://github.com/odoo/enterprise/blob/313593cf176271060878dda5ebc677fa44b519db/partner_commission/models/account_move.py#L46-L50

opw-4676967

closes odoo/enterprise#83725

X-original-commit: ae710c83995c625ecc1e4ed6da8857f1431e7321
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>