[FIX] sale_purchase_inter_company_rules: check sales team company on Inter-Company sync
Steps to reproduce:

- With Company B in the settings Enable Inter-Company Transactions
   > Synchronize Sales and Purchase Order
- Sales > Configurations > Sales Teams, Archive all sales team
- With COMP A create a sales team with a set company_id: COMP A
- With COMP A create and confirm a PO with COMP B as customer
> Invalid operation: while generating the SO, the sales team belongs to
  COMP A and the SO to COMP B.

Cause of the issue:

The `team_id` field of the `sale.order` model is a stored precomputed
and company checked field. As such, when the SO is created and even if
no sales team is provided to the create vals:
https://github.com/odoo/enterprise/blob/c61ce5a8e46e706bfb2025d4c9c79a39598e8827/sale_purchase_inter_company_rules/models/purchase_order.py#L70
a default sales team will be computed and set on the SO based on the
"allowed_company_ids" (including both COMP A and COMP B):
https://github.com/odoo/odoo/blob/1fa8678a33ab35195005d1b65b5eecb1d089fe56/addons/sale/models/sale_order.py#L453-L460
https://github.com/odoo/odoo/blob/1fa8678a33ab35195005d1b65b5eecb1d089fe56/addons/sales_team/models/crm_team.py#L81-L88
While the domain checks that the sales team is either not tight to a
company or belong to comp B, since no such teams were found, we fallback
the invalid sales team of COMP A. The invalid operation is then raised
during the `_check_company` of the created record.

opw-4627851

closes odoo/enterprise#87465

X-original-commit: be3961c81c6a9bcb8703bd8d0a2b84a82e942457
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>