[FIX] knowledge: upgrade searchModel favorite filters `user_id` field
[This commit] modified the `ir.filters` `user_id` field into `user_ids`.
In Knowledge embedded views (and links), the `SearchModel` state and some
custom favorite filters are stored inside a html attribute of the body of an
article, therefore they are not properly upgraded with the standard python
script.

This commit introduces a JS upgrade script to modify these values. `user_id` is
converted to `user_ids` with an empty array as a value (which means that the
filters are always "shared with everyone").

The field has to be modified both in the custom favorites and in the stored
`SearchModel` state.

[This commit]: https://github.com/odoo/odoo/commit/414e55cf7c3971a7ba0a7c96855db4c23e77a5c4

task-4718129

closes odoo/enterprise#85112

X-original-commit: 1dd6d92c269a342b5df898d5bb2a9efe3ec21222
Related: odoo/odoo#209239
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>