[FIX] l10n_mx_edi: CRP20211 using foreign currency bank transaction
Steps to reproduce:
(thanks to TBS for investigating + writing the test)

1/ Have a MX company with Quadrum or SwSapien
2/ Create a USD bank transaction with 13220.00 USD and 305147.51 MXN
3/ Create a USD invoice for 13220.00 USD and send CFDI
4/ Reconcile transaction and invoice
5/ Click "Update payments" on the invoice => CRP20211

Looks like it comes from this rounding difference

Rate is computed by odoo using
305147.51 MXN / 13220.0 USD = 23.082262481 ≃ 23.082262 MXN/USD
PAC verify the amount paid using
13220.0 USD * 23.082262 MXN/USD = 305147.50 MXN which is not 305147.51 MXN

Solution:
There is no solution by adapting the rate:
- quadrum restricts the rate to 6 decimals;
- rounding the rate differently (e.g. to 23.082263) would not help
  because 13220.0 USD * 23.082263 MXN/USD = 305147.52 MXN which again is
  not 305147.51 MXN.

So, the only solution that I can think of is to adapt the amount in
company currency so that it is exactly
round(amount_in_payment_curr / rate_rounded_to_6_decimals),
i.e. in the example, change it to 305147.50 MXN.

In tests, both Quadrum and SwSapien seem to accept this.

We don't perform the modification for other PACs.

opw-4222310

closes odoo/enterprise#86074

X-original-commit: b54905d99c04529c8a60ba489e1f39ec8e444130
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Co-authored-by: Thomas Beckers <tbs@odoo.com>