[FIX] web_studio: hide 'Add a button' from subview editor
This commit fixes the implementation of the 'Add a button' feature when editing
a subview like the x2many list. Neither the parameters nor the python function
support adding a button in a subview. This feature would in itself makes little
sense as records cannot be selected in this case.

closes odoo/enterprise#80198

Related: odoo/odoo#199373
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>