[IMP] voip: merge CallView into UserInfo
CallView was introduced to avoid code duplication between
CallInvitation, InCallView, and CallSummary, which have some data in
common.

The concept of CallView is no longer useful since the part that is
shared is fully encompassed by the UserInfo component that was added
later. This commit merges the two to keep only one component.

Part-of: odoo/enterprise#83544
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>