[FIX] hr_payroll: fixes color isn't updating
Steps to Produce:
open payroll app
Go to Salary Rules or go to Salary Structures and open any salary rule.
try to change color of any rule.

issue:
The color does not change in the Salary Rules.
Additionally, styles do not apply in the salary rule pop-up accessed through
Salary Structures.

Fix:
Instead of changing the color with formrenderer, a custom widget was
created to apply styles directly to the field.

opw-4716810
task-4774448

closes odoo/enterprise#93608

X-original-commit: c78a21e7bfd176b386b45751fcb77283df90e128
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>