[FIX] sendcloud_service: Prevent division by zero for average cost
Beofre this commit:

Products with both positive and negative sales order lines in SO can result it total quantity = 0 when calcuating average cost.

Causing division by zero error.

After this commit:

Added a check to prevent division if total quantity of product is zero.

Negative order lines are usually used for return of product, it is not included in shipping request.

opw-4655713

closes odoo/enterprise#84821

X-original-commit: 7495abd08889001af79fc1f241dc34f80c51ed59
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Basit Arshad (baar) <baar@odoo.com>