[FIX] account_reports: fix adjustment entry in unrealized currency gains report
Steps to reproduce:
- Install Accounting

- Activate a foreign currency (e.g. EUR)
- Make sure there are several rates with today rate different than others
- Go to "Accounting / Configuration / Accounting / Chart of Accounts"
- Add a default tax to "Expenses" account

- Create an invoice:
  * Customer: [any]
  * Currency: [EUR] (i.e. the foreign currency)
  * Invoice Date: [a date in the past where the currency rate is different than today]
  * Invoice Lines: [any]
- Confirm the invoice

- Go to "Accounting / Reporting / Management / Unrealized Currency Gains/Losses"
- Set date filter to "Today"
- If there is no adjustement to do, check if the following message is displayed:
"You are using custom exchange rates. => Reset to Odoo’s Rate"
If it's the case, reset to Odoo's rate (i.e. Today's rate)
- Click on "Adjustement Entry" button:
  * Journal: Miscellaneous Operations
  * Expense Account: Expenses
  * Income Account: [any]
- Create Entry

Issue:
The values (debit/credit) of the created entry are 0, which is not the case
when there is no default tax on the expense account.

Cause:
"_sync_dynamic_lines" is executed on the created entry.
If a default tax is set on the expense account, "_sync_tax_lines" is also
executed.
As only debit and credit values are set on the invoice lines of the entry
(neither balance nor amount_currency), the sync method messes up with the
value of debit and credit, resetting them to 0 (via multiple recompute).

Solution:
Skip invoice sync.
The values computed in "Multicurrency Revaluation Wizard" should be the
ones that are used.

opw-4664255

closes odoo/enterprise#83993

X-original-commit: d89ec3d548a55d8abd7df1b24309b0042eb99e8c
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>