[FIX] l10n_de_pos_res_cert: Correctly handle print error
When printer isn't reacheable a traceback is shown when trying to
print a receipt. This is because the error is not handled correctly.

This commit adds a check to ensure that the error is correctly
handled and that the user is notified of the error in a user-friendly
manner.

closes odoo/enterprise#83611

X-original-commit: a34df2f14f7237612a8d40e991cf4980c1e37ee1
Related: odoo/odoo#206457
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: David Monnom (moda) <moda@odoo.com>