[FIX] l10n_mx_edi: ignore bin_size when uploading edi xml to read it correctly
**Steps to reproduce:**
With mexican localisation installed:
- Open form view for the model 'l10n_mx_edi.document'
- Prepare a payment document in the Payment-20 format (https://www.sat.gob.mx/sitio_internet/cfd/Pagos/Pagos20.xsd)
- Remove existing attachment
- Change the attachment on the document to the relatively long XML file

**Issue:**
After saving the change, the attachment_uuid (Fiscal Folio) is left blank.

**Cause:**
Normally, the uuid should be extracted from the uploaded attachment, but
the computation of the 'raw' field on the attachment returns the value
"b'56.00 bytes'", leaving the attachment_uuid field empty because the XML
is assumed to be in the wrong format.
The issue doesn't exist for shorter XML files in the Payment-10 format
(https://www.sat.gob.mx/sitio_internet/cfd/Pagos/Pagos10.xsd)
**Solution:**
Before computing the attachment_uuid from the attachment, we set the
bin_size to False in the context.

opw-4641487

closes odoo/enterprise#86125

X-original-commit: 242aa31dc1b0ff7b8161cd889a305e45249ea088
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>