[FIX] (website_)knowledge: search bar with incorrect domain
This commit fixes an issue with the KnowledgeSidebar Component which provides a
domain to search for a specific article. The issue is that it used the field
`website_published` inside the knowledge module. Which never defines this field
on the model. This error was introduced by odoo/enterprise#77597.

This is supposed to be the job of the website_knowledge module, which implements
the field on the model.

Another issue is when a portal user access a published
article and switches to an article shared with them, he can not search for
articles that are descendants of the original published article.

This is fixed by registering the original root id (which may be a
published article), and searching published articles that are children
of that specific root.

task-4734644

closes odoo/enterprise#83560

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>