[FIX] social_youtube: prevent traceback on video upload
When trying to upload the video for a Youtube post, we are getting a
traceback with "UncaughtPromiseError > TypeError", caused by not having
the `youtube_video_category_id` field properly loaded, at the time where
we are trying to evaluate `_preprocessReferenceChanges` we try to access
the type of the field as `this.fields[fieldName].type === "reference"`
where `fieldName` is `youtube_video_category_id`, but the field is not
loaded yet (this.fields[fieldName] is undefined), so we get an error.

This commit adds this field as a field dependecy for the
YoutubeUploadField widget, so that it is loaded when we try to access it.

opw-4646432

closes odoo/enterprise#87151

X-original-commit: 96b10b97b0553255a538a4c3ea5ce0bb9aa8ac8b
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>