[IMP] whatsapp: increase WhatsApp API request timeout
When submitting message templates to the WhatsApp API, the request could
occasionally time out before receiving a response. This resulted in
templates being created on Facebook's side, but not registered in Odoo,
as the template ID assigned by Facebook was never received.

Since template names must be unique, resubmitting the same template
fails, leaving the systems out of sync.

This commit increases the request timeout to (10s connect, 30s read) to
reduce the chance of timeouts.

opw-4706852
opw-4720635

closes odoo/enterprise#84343

X-original-commit: 63aa90487c9a3aa39eccfe312494f990138907b4
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>