[FIX] stock_barcode: avoid demand updates on neutral change
Issue: In Barcode, editing the qty_done of a move line and then reseting
it to 0 will update the demand.

Steps to reproduce:

- Create a storable product. Put 5 units in stock.
- Create a delivery order for 5 units and mark as to do.
- Go to the barcode app and edit your delivery order.
- Click on the pencil and set a qty of 1 > Confirm
- Click back on the pencil and reset the quantity to 0 > confirm
> On the picking, the demand of the line was edited from 5 to 1.

Cause of the issue:

Editing the quantity of a move line from the barcode actually updates
the qty_done of the move line which updates both the quantity and the
picked status of the move line in the backend via the `_inverse_qty_done`:
https://github.com/odoo/enterprise/blob/cbcff76a94750a6759fd72d9c952d785e603a974/stock_barcode/models/stock_move_line.py#L58-L62
Initially, the quantity of the move line was 5 and hence the demand of
the line in the barcode was displayed as 5:
https://github.com/odoo/enterprise/blob/cbcff76a94750a6759fd72d9c952d785e603a974/stock_barcode/static/src/models/barcode_picking_model.js#L119-L121
The first time you edit the line, you will actually change its quantity
to 1 and set it as picked. Once the line state refreshes in the barcode
it will then keep the same demand of 5 as it will then be based on the
previous demand of the line (since the qty of the move line is set
and the qty done is positive):
https://github.com/odoo/enterprise/blob/cbcff76a94750a6759fd72d9c952d785e603a974/stock_barcode/static/src/models/barcode_picking_model.js#L1037-L1045
However, the next time you update the line the `_inverse_qty_done` will
leave the quantity of the move line at 1 but will also unpick it. Then,
and since the `qty_done` was set to 0, it will update the demand of the
barcode line because of these same lines:
https://github.com/odoo/enterprise/blob/cbcff76a94750a6759fd72d9c952d785e603a974/stock_barcode/static/src/models/barcode_picking_model.js#L1037-L1045

Fix:

Prior to Commit 3af3e00aabecde8b811ac2cbab397f982721f164 the quantity
was updated as well as the picked state even when the qty_done was set
to 0, so that this issue would not occur as the null quantity would
make us keep the demand of barcode line after edition. However, the
problem was that this would update the reserved quantity of the move
and hence the barcode app demand once you leave the barcode app because
of the `split_uncompleted_moves`. However, in the present use case,
that same method can save us because if the move is picked and its
quantity is null when you exit the barcode app, it will be rereversed.

opw-4700996

closes odoo/enterprise#87047

X-original-commit: 5e2dfb22494d39c64fa2fd6ebd0679ebadff9e10
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>