[FIX] l10n_it_xml_export: LIPE export tracebacks
Linked to this fix: https://github.com/odoo/enterprise/commit/7323435468c8db15cd5643f7d6ecf5928ab2f3dc
During the fix linked above, we thought 17.0 didn't need a fix, it does. The
behavior needs to be adapted depending on whether we are on the report or on
the closing entry.

This fix makes sure the correct method is chosen to gather the options required
for the xml export of the tax return.

no-task

X-original-commit: 9b61a7a87c192380b2b080261e24f623bc731d6e
Part-of: odoo/enterprise#83744
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>