[REF] knowledge: no longer use "onWillStart" model option
This hook has been removed in odoo/odoo#205129, as it was only used
by the knowledge calendar view, and using the hook wasn't necessary.
What was done in it doesn't involve loading data, it was only about
setupping the model. So this can simply be done in the setup, after
calling super.

Part of task-205129

closes odoo/enterprise#83321

Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>