[FIX] pos_blackbox_be: fix user form xpath after UI redesign
The pos_blackbox_be module was broken in master because PR #190466 rewrote the
UI of the users form, causing the xpath used in the view inheritance to fail. This
commit updates the xpath to work with the new UI structure, ensuring that the INSZ
number field displays correctly in the user form.

closes odoo/enterprise#83669

Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>