[FIX] pos_settle_due: fix test
The same test is used in multiple modules. Sometimes in the mx
localization it was possible that the closing of the session
would take a bit longer and the modal wasn't fully closed (though
invisible.

The modal step in the tour was added when it was planned to
remove the 'in_modal' parameter in tour steps and is not really
needed for the original purpose of the test.

closes odoo/enterprise#83855

Runbot-error: 163628
X-original-commit: 25dc5118beb43fe2e03ad664462ac4b47898b283
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>