[FIX] l10n_it_xml_export: resume move post after export
Description of the issue this commit addresses:

When posting the closing entry of the italian periodic tax report, the wizard
for xml export opens and stops the posting process so when the file is
exported, the moves are actually not posted.

Desired behavior after this commit is merged:

As late as possible before exporting the xml file, the closing entry move is
posted.

no-task

closes odoo/enterprise#83744

X-original-commit: 5001fd1df3237dd88629fed5b0f6df3fa863d648
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>