[IMP] knowledge: load property fields in edit mode by default
In the community PR, we introduced an edit mode for property fields.
Because the Knowledge editor doesn't include a cog menu to toggle this
mode, the property field component now defaults to edit mode — provided
the user has "write" access to the definition record.

To reduce visual clutter, the buttons for editing or moving a property
field will be hidden by default and shown only on hover. This helps
create a cleaner and more intuitive interface.

See odoo/odoo#203938

Task-4610816

Part-of: odoo/enterprise#82574
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>