[IMP] hr_payroll: move credit time from standard to BE
The credit time mechanism is a Belgian legal thing.
In Belgium, instead of having a part time parameter on the version, we have now a parameter on the work entry type to specify if is it credit time or not.
To create a credit time version, we should specify in the "Working schedule change" wizard, a working schedule with lines with a credit time work entry type.

task-4578909

closes odoo/enterprise#83375

Related: odoo/odoo#206658
Related: odoo/upgrade#8158
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>