[FIX] iot: call discover boxes from client
Before this commit, the `/iot/discover-boxes` endpoint was called from
the backend. This meant that the associated IP was that of the DB, which
is usually different from the client network and not the intended
behaviour.

After this commit, the discover boxes route is called in a client
action, and the result passed to the backend. There is a new transient
model to keep track of the discovered IoT boxes during the wizard's
lifetime, which simplifies the selection logic.

task-4687795

closes odoo/enterprise#83726

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>